+++ B2T2_Alvarez/B2T2_Alvarez/EBO.h
@@ -7,7 +7,7 @@ class EBO
 {
 public:
 	GLuint ID;
-	EBO(GLfloat* vertices, GLsizeiptr size);
+	EBO(GLuint* vertices, GLsizeiptr size);
 
 	void Bind();
 	void Unbind();
